Consolidate short files and remove re-export facades
- Merge scip-symbols.ts + scip-types.ts into single scip.ts (semantic/symbols)
- Remove line-comment/utils.ts re-export facade; update consumers to import directly
- Merge string-coercions.ts into params.ts (sole consumer)
- Remove doc-comment/utils.ts re-export facade; update consumers to import directly
- Update all affected imports and tests

Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/feea2025-2e14-4696-a24f-fcfccdeceaeb
