+++ src/core/src/comments/doc-comment/params.ts
@@ -1,4 +1,12 @@
-import { defaultDocCommentStringCoercions, type DocCommentStringCoercions } from "./string-coercions.js";
+import { getNonEmptyString } from "../../utils/string.js";
+
+export interface DocCommentStringCoercions {
+    coerceNonEmptyString(value: string): string | null;
+}
+
+export const defaultDocCommentStringCoercions: DocCommentStringCoercions = Object.freeze({
+    coerceNonEmptyString: getNonEmptyString
+});
 
 const STRING_TYPE = "string";
