--- src/core/src/comments/doc-comment/string-coercions.ts
@@ -1,9 +1,0 @@
-import { getNonEmptyString } from "../../utils/string.js";
-
-export interface DocCommentStringCoercions {
-    coerceNonEmptyString(value: string): string | null;
-}
-
-export const defaultDocCommentStringCoercions: DocCommentStringCoercions = Object.freeze({
-    coerceNonEmptyString: getNonEmptyString
-});
