+++ src/semantic/src/scopes/scope-tracker.ts
@@ -1,6 +1,6 @@
 import { Core, type MutableGameMakerAstNode } from "@gmloop/core";
 
-import { ROLE_DEF, ROLE_REF } from "../symbols/scip-types.js";
+import { ROLE_DEF, ROLE_REF } from "../symbols/scip.js";
 import { IdentifierCacheManager } from "./identifier-cache-manager.js";
 import { cloneDeclarationMetadata, cloneOccurrence, createOccurrence } from "./occurrence.js";
 import { GlobalIdentifierRegistry } from "./registry.js";
