+++ src/semantic/src/symbols/index.ts
@@ -1,4 +1,3 @@
 export * from "./built-in-identifiers.js";
-export * from "./scip-symbols.js";
-export * from "./scip-types.js";
+export * from "./scip.js";
 export * from "./sem-oracle.js";
