--- src/semantic/src/symbols/scip-symbols.ts
@@ -1,14 +1,0 @@
-/**
- * @typedef {"script"|"event"|"object"|"macro"|"enum"|"var"} GmlSymbolKind
- */
-
-/**
- * Create a stable SCIP symbol string for a GML symbol.
- * @param {GmlSymbolKind} kind
- * @param {string} name
- */
-export function sym(kind, name) {
-    return `gml/${kind}/${name}`;
-}
-
-export default { sym };
