--- src/semantic/src/symbols/scip-types.ts
@@ -1,10 +1,0 @@
-// Range tuple: [startLine, startCol, endLine, endCol]
-export const ROLE_DEF = 1;
-export const ROLE_REF = 0;
-
-// Runtime exports only; structural SCIP types live in ../scopes/types.ts.
-
-export default {
-    ROLE_DEF,
-    ROLE_REF
-};
