+++ src/semantic/test/scope-tracker-scip-export.test.ts
@@ -2,7 +2,7 @@ import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 
 import { ScopeTracker } from "../src/scopes/scope-tracker.js";
-import { ROLE_DEF, ROLE_REF } from "../src/symbols/scip-types.js";
+import { ROLE_DEF, ROLE_REF } from "../src/symbols/scip.js";
 
 const createScriptSymbol = (name: string): string => `gml/script/${name}`;
